Prevent truncated text results in previews
We clip the search regions  a little greater than the visible region.
We might get a truncated color #ABC which is actually #ABCDEF.  By
requiring the matched results to be entirely in the visible region, we
can ensure we had enough context after the color to verify the match is
good.