Prevent duplicate previews
If somehow we have two entries that have phantoms that start at the
same point, something went wrong.  It is likely things got out of sync
and one phantom was created and a different location than the second,
but the phantoms reference the same color.  We will lose track of all
duplicates as we will update the preview dictionary with where the
previews location is NOW, not where they were originally created.  When
we do that, the key will be the same for all duplicates.  This could
cause us to keep overwriting the same key when rebuilding the preview
after erasure; we would lose track of all duplicates but one and think
there is only one.

When rebuilding the preview list, don’t append if the point is already
filled.

fixes #57 (hopefully)