Another weird Sublime view settings case
Another situation where Sublime returns None from the settings
dictionary via "get" when it should return the default or the actual
value.  This seems to be some kind of race condition as we never set the
object in the settings file None, and this isn't the first time I've
seen this.

In this case, I am not certain if the object is already in the view or
not, but if it isn't, we should have used the default value of "{}", and
if it was, we should have pulled the dictionary object, but isntead we
got None which we never set.